refactor: legacy miss count calc
